# controllers/clientesController.js
const ClientesModel = require("../models/clientesModel");
const bcrypt = require("bcryptjs");
const jwt = require("jsonwebtoken");
const keys = require("../config/keys");
const { sendValidationError, validatePagination } = require("../validators/common");
const { validateCreateCliente, validateUpdateCliente } = require("../validators/clientes");

const ClientesController = {
  /**
   * @swagger
   * components:
   *   schemas:
   *     Cliente:
   *       type: object
   *       properties:
   *         idClientes:
   *           type: integer
   *         nomeCliente:
   *           type: string
   *         email:
   *           type: string
   *         senha:
   *           type: string
   */

  /**
   * @swagger
   * /clientes:
   *   get:
   *     summary: Obtiene la lista de clientes
   *     responses:
   *       200:
   *         description: Lista de clientes
   *         content:
   *           application/json:
   *             schema:
   *               type: array
   *               items:
   *                 $ref: '#/components/schemas/Cliente'
   */
  getClientes: (req, res) => {
    const { id } = req.query;
    const pagination = validatePagination(req.query);

    if (pagination.errors.length > 0) {
      return sendValidationError(res, pagination.errors);
    }

    if (id) {
      ClientesModel.getById(id, (err, cliente) => {
        if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });
        if (!cliente)
          return res.status(404).send({ message: "Cliente no encontrado" });

        ClientesModel.getAllOsByClient(id, (err, os) => {
          if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });
          cliente.ordensServicos = os;
          res
            .status(200)
            .send({ message: "Detalles del Cliente", result: cliente });
        });
      });
    } else {
      // Pass search term directly - model handles parameterization (prevents SQL injection)
      ClientesModel.get(
        "clientes",
        "*",
        pagination.search,
        pagination.perPage,
        pagination.start,
        (err, clientes) => {
          if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });
          if (!clientes || clientes.length === 0)
            return res
              .status(404)
              .send({ message: "Ningún cliente localizado." });
          res
            .status(200)
            .send({ message: "Lista de Clientes", result: clientes });
        }
      );
    }
  },

  /**
   * @swagger
   * /clientes:
   *   post:
   *     summary: Crea un nuevo cliente
   *     requestBody:
   *       required: true
   *       content:
   *         application/json:
   *           schema:
   *             $ref: '#/components/schemas/Cliente'
   *     responses:
   *       201:
   *         description: Cliente añadido con éxito
   *       500:
   *         description: Error en el servidor
   */
  createCliente: (req, res) => {
    const validation = validateCreateCliente(req.body || {});

    if (validation.errors.length > 0) {
      return sendValidationError(res, validation.errors);
    }

    const data = { ...validation.data };
    const cpf_cnpj = String(data.documento).replace(/[^\w\s]/gi, "");
    const senhaCliente = data.senha || cpf_cnpj;
    data.senha = bcrypt.hashSync(senhaCliente, 10);
    data.pessoa_fisica = cpf_cnpj.length === 11;
    data.dataCadastro = new Date();
    data.fornecedor = data.fornecedor ? 1 : 0;

    ClientesModel.add("clientes", data, (err, result) => {
      if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });
      ClientesModel.getById(result, (err, cliente) => {
        if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });
        res
          .status(201)
          .send({ message: "Cliente añadido con éxito", result: cliente });
      });
    });
  },

  /**
   * @swagger
   * /clientes/{id}:
   *   put:
   *     summary: Actualiza un cliente existente
   *     parameters:
   *       - in: path
   *         name: id
   *         required: true
   *         schema:
   *           type: integer
   *     requestBody:
   *       required: true
   *       content:
   *         application/json:
   *           schema:
   *             $ref: '#/components/schemas/Cliente'
   *     responses:
   *       200:
   *         description: Cliente editado con éxito
   *       404:
   *         description: Cliente no encontrado
   *       500:
   *         description: Error en el servidor
   */
  updateCliente: (req, res) => {
    const { id } = req.params;
    const validation = validateUpdateCliente(req.body || {});

    if (validation.errors.length > 0) {
      return sendValidationError(res, validation.errors);
    }

    const data = { ...validation.data };

    if (data.senha) {
      data.senha = bcrypt.hashSync(data.senha, 10);
    }

    ClientesModel.edit("clientes", data, "idClientes", id, (err, result) => {
      if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });
      if (result.affectedRows === 0)
        return res.status(404).send({ message: "Cliente no encontrado" });
      ClientesModel.getById(id, (err, cliente) => {
        if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });
        res
          .status(200)
          .send({ message: "Cliente editado con éxito", result: cliente });
      });
    });
  },

  /**
   * @swagger
   * /clientes/{id}:
   *   delete:
   *     summary: Elimina un cliente existente
   *     parameters:
   *       - in: path
   *         name: id
   *         required: true
   *         schema:
   *           type: integer
   *     responses:
   *       200:
   *         description: Cliente eliminado con éxito
   *       404:
   *         description: Cliente no encontrado
   *       500:
   *         description: Error en el servidor
   */
  deleteCliente: (req, res) => {
    const { id } = req.params;

    ClientesModel.getAllOsByClient(id, (err, os) => {
      if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });
      const osIds = os.map((o) => o.idOs);
      ClientesModel.removeClientOs(osIds, (err, result) => {
        if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });

        ClientesModel.getAllVendasByClient(id, (err, vendas) => {
          if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });
          const vendaIds = vendas.map((v) => v.idVendas);
          ClientesModel.removeClientVendas(vendaIds, (err, result) => {
            if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });

            ClientesModel.delete(
              "clientes",
              "idClientes",
              id,
              (err, result) => {
                if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });
                if (result.affectedRows === 0)
                  return res
                    .status(404)
                    .send({ message: "Cliente no encontrado" });
                res
                  .status(200)
                  .send({ message: "Cliente eliminado con éxito" });
              }
            );
          });
        });
      });
    });
  },

  /**
   * @swagger
   * /clientes/login:
   *   post:
   *     summary: Autentica un cliente
   *     requestBody:
   *       required: true
   *       content:
   *         application/json:
   *           schema:
   *             type: object
   *             properties:
   *               email:
   *                 type: string
   *               password:
   *                 type: string
   *     responses:
   *       200:
   *         description: Login exitoso
   *       401:
   *         description: Credenciales incorrectas
   *       500:
   *         description: Error en el servidor
   */
  login: (req, res) => {
    const { email, password } = req.body;

    if (!email || !password) {
      return res.status(400).send({ message: "Email y contraseña son requeridos" });
    }

    ClientesModel.getClienteByEmail(email, (err, cliente) => {
      if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });

      if (
        !cliente ||
        typeof password !== "string" ||
        typeof cliente.senha !== "string" ||
        !bcrypt.compareSync(password, cliente.senha)
      ) {
        return res.status(401).send({ message: "Credenciales incorrectas" });
      }

      const token = jwt.sign(
        { id: cliente.idClientes, email: cliente.email, rol: "cliente" },
        keys.secretOrKey,
        {
          expiresIn: keys.expiresIn || "1h",
        }
      );

      res.status(200).send({ 
        message: "Login exitoso", 
        token,
        cliente: {
          id: cliente.idClientes,
          nomeCliente: cliente.nomeCliente,
          email: cliente.email,
          rol: "cliente",
        },
      });
    });
  },

  /**
   * @swagger
   * /clientes/os/{id}:
   *   get:
   *     summary: Obtiene las órdenes de servicio de un cliente por ID
   *     parameters:
   *       - in: path
   *         name: id
   *         required: true
   *         schema:
   *           type: integer
   *     responses:
   *       200:
   *         description: Ordenes de servicio
   *         content:
   *           application/json:
   *             schema:
   *               type: array
   *               items:
   *                 type: object
   *                 properties:
   *                   idOs:
   *                     type: integer
   *                   descricaoProduto:
   *                     type: string
   *                   defeito:
   *                     type: string
   *                   observacoes:
   *                     type: string
   *                   dataInicial:
   *                     type: string
   *                     format: date-time
   *                   dataFinal:
   *                     type: string
   *                     format: date-time
   *                   valorTotal:
   *                     type: number
   *                     format: float
   *       404:
   *         description: Cliente no encontrado
   *       500:
   *         description: Error en el servidor
   */
  getOsByIdClientes: (req, res) => {
    const { id } = req.params;
    ClientesModel.getAllOsByClient(id, (err, os) => {
      if (err) return res.status(500).send({ message: "Error del servidor", error: err.message });
      res.status(200).send({ message: "Ordenes de servicio", result: os });
    });
  },

  /**
   * @swagger
   * /clientes/compras/{id}:
   *   get:
   *     summary: Obtiene todas las compras de un cliente por ID
   *     parameters:
   *       - in: path
   *         name: id
   *         required: true
   *         schema:
   *           type: integer
   *     responses:
   *       200:
   *         description: Compras del cliente
   *         content:
   *           application/json:
   *             schema:
   *               type: array
   *               items:
   *                 type: object
   *                 properties:
   *                   idCompra:
   *                     type: integer
   *                   descricaoProduto:
   *                     type: string
   *                   dataCompra:
   *                     type: string
   *                     format: date-time
   *                   valorTotal:
   *                     type: number
   *                     format: float
   *       404:
   *         description: Cliente no encontrado
   *       500:
   *         description: Error en el servidor
   */
  getAllComprasByClientes_id: (req, res) => {
    const { id } = req.params;
    ClientesModel.getAllComprasByClientes_id(id, (err, compras) => {
      if (err) return res.status(500).send({ message: "Error del servidor", error: err.message }); // Fixed: was res.status(500(err))
      res.status(200).send({ message: "Compras del cliente", result: compras });
    });
  },

  /**
   * @swagger
   * /clientes/cobranzas/{id}:
   *   get:
   *     summary: Obtiene todas las cobranzas de un cliente por ID
   *     parameters:
   *       - in: path
   *         name: id
   *         required: true
   *         schema:
   *           type: integer
   *     responses:
   *       200:
   *         description: Cobranzas del cliente
   *         content:
   *           application/json:
   *             schema:
   *               type: array
   *               items:
   *                 type: object
   *                 properties:
   *                   idCobranza:
   *                     type: integer
   *                   descricaoCobranza:
   *                     type: string
   *                   dataCobranza:
   *                     type: string
   *                     format: date-time
   *                   valorTotal:
   *                     type: number
   *                     format: float
   *       404:
   *         description: Cliente no encontrado
   *       500:
   *         description: Error en el servidor
   */
  getAllCobranzasByClientes_id: (req, res) => {
    const { id } = req.params;
    ClientesModel.getAllCobranzasByClientes_id(id, (err, cobranzas) => {
      if (err) return res.status(500).send({ message: "Error del servidor", error: err.message }); // Fixed: was res.status(500(err))
      res
        .status(200)
        .send({ message: "Cobranzas del cliente", result: cobranzas });
    });
  },
};

module.exports = ClientesController;