# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentación Taller Digicom By: Richard Ortiz",
    "version": "1.0.0",
    "description": "Documentación de la API para la gestión de talleres"
  },
  "servers": [
    {
      "url": "https://api.taller.digicom.com.gt/api/v1",
      "description": "Production API over HTTPS"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development API"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Cliente": {
        "type": "object",
        "properties": {
          "idClientes": {
            "type": "integer"
          },
          "nomeCliente": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/clientes": {
      "get": {
        "summary": "Obtiene la lista de clientes",
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer"
            },
            "default": 20
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idClientes": {
                        "type": "integer"
                      },
                      "nomeCliente": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      }
                    },
                    "$ref": "#/components/schemas/Cliente"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ningún cliente localizado."
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      },
      "post": {
        "summary": "Crea un nuevo cliente",
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nomeCliente": {
                    "type": "string"
                  },
                  "documento": {
                    "type": "string"
                  },
                  "senha": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "$ref": "#/components/schemas/Cliente"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente añadido con éxito"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/clientes/{id}": {
      "get": {
        "summary": "Obtiene los detalles de un cliente por ID",
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del Cliente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "idClientes": {
                      "type": "integer"
                    },
                    "nomeCliente": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "ordensServicos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "idOs": {
                            "type": "integer"
                          },
                          "descricaoProduto": {
                            "type": "string"
                          },
                          "dataCadastro": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "valorTotal": {
                            "type": "number",
                            "format": "float"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cliente no encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      },
      "put": {
        "summary": "Actualiza un cliente existente",
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nomeCliente": {
                    "type": "string"
                  },
                  "documento": {
                    "type": "string"
                  },
                  "senha": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "$ref": "#/components/schemas/Cliente"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente editado con éxito"
          },
          "404": {
            "description": "Cliente no encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      },
      "delete": {
        "summary": "Elimina un cliente existente",
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente eliminado con éxito"
          },
          "404": {
            "description": "Cliente no encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/clientes/os/{id}": {
      "get": {
        "summary": "Obtiene las órdenes de servicio de un cliente por ID",
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordenes de servicio",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idOs": {
                        "type": "integer"
                      },
                      "descricaoProduto": {
                        "type": "string"
                      },
                      "dataCadastro": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "valorTotal": {
                        "type": "number",
                        "format": "float"
                      },
                      "defeito": {
                        "type": "string"
                      },
                      "observacoes": {
                        "type": "string"
                      },
                      "dataInicial": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "dataFinal": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cliente no encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/clientes/compras/{id}": {
      "get": {
        "summary": "Obtiene todas las compras de un cliente por ID",
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compras del cliente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idCompra": {
                        "type": "integer"
                      },
                      "descricaoProduto": {
                        "type": "string"
                      },
                      "dataCadastro": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "valorTotal": {
                        "type": "number",
                        "format": "float"
                      },
                      "dataCompra": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cliente no encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/clientes/cobranzas/{id}": {
      "get": {
        "summary": "Obtiene todas las cobranzas de un cliente por ID",
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cobranzas del cliente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idCobranca": {
                        "type": "integer"
                      },
                      "descricaoProduto": {
                        "type": "string"
                      },
                      "dataCadastro": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "valorTotal": {
                        "type": "number",
                        "format": "float"
                      },
                      "idCobranza": {
                        "type": "integer"
                      },
                      "descricaoCobranza": {
                        "type": "string"
                      },
                      "dataCobranza": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cliente no encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/clientes/login": {
      "post": {
        "summary": "Autentica un cliente",
        "tags": [
          "Clientes"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso"
          },
          "401": {
            "description": "Credenciales incorrectas"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
...
  },
...
}